# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Min Khaung Tea House & Restaurant Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a new table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get a table by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "404":
          description: Not found
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update table status/info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableBody"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── MENU ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: integer
        tableNumber:
          type: string
        zone:
          type: string
          description: "hall or aircon"
        capacity:
          type: integer
        status:
          type: string
          description: "available, occupied, payment_pending, dirty"
        qrCode:
          type: string
        posX:
          type: integer
        posY:
          type: integer
        currentOrderId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tableNumber
        - zone
        - capacity
        - status
        - posX
        - posY
        - createdAt
        - updatedAt
    CreateTableBody:
      type: object
      properties:
        tableNumber:
          type: string
        zone:
          type: string
        capacity:
          type: integer
        posX:
          type: integer
        posY:
          type: integer
      required:
        - tableNumber
        - zone
        - capacity
        - posX
        - posY
    UpdateTableBody:
      type: object
      properties:
        tableNumber:
          type: string
        zone:
          type: string
        capacity:
          type: integer
        status:
          type: string
        posX:
          type: integer
        posY:
          type: integer
        currentOrderId:
          type: ["integer", "null"]

    # ─── MENU SCHEMAS ─────────────────────────────────────────────────────
...