# lib/api-zod/src/generated/types/updateTableBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Min Khaung Tea House & Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTableBody {
  tableNumber?: string;
  zone?: string;
  capacity?: number;
  status?: string;
  posX?: number;
  posY?: number;
  /** @nullable */
  currentOrderId?: number | null;
}